Create readme.md
The research examines the application of concepts from evolutionary game theory’s replicator dynamics and their ability to optimize learning, stability, and adaptation in Multi-Agent Reinforcement learning (MARL). Multi-Agent Reinforcement learning involves adapting the standard reinforcement learning framework to work in a situation where there are multiple learning agents operating in the same environment. This allows the environment to be dynamically shifting because the various agents can be cooperating, competing, or even doing both.

Replicator dynamics: This classical model of evolutionary game theory captures the evolution of strategies based on their relative payoffs. A strategy will grow if it outperforms the average payoff in the population and decay otherwise. This model has been incorporated in MARL as an efficient learning direction towards fixed points of the game dynamics such as Nash equilibria or evolutionarily stable strategies. Previous research has demonstrated the effectiveness of models like Replicator Equations over policy gradient methods in non-stationary and adversarial environments. This involves the NeuRD method of modifying policy gradients to be similar to Replicator Equations to result in quicker and more stable convergence to the Nash Equilibrium of poker games. There has also been research highlighting the capability of models involving Replicator Equations combined with environment information that can generate complex fixed points such as stable points and cycles.

Based on the above concepts, the approach combines two parts: the update rules inspired from NeuRD and the revision rules from ERI-D. These rules selectively revise the experiences in the replay buffer. A parameter determines the extent to which the approach follows the evolutionary update rules rather than the standard gradient update rules. The model makes adjustments according to the adapting opponent behaviors and the varying value of the reward. The state-coupled replicator dynamics allow the model to be stable when there are sudden transitions and/or sudden changes in the value of the reward.

This approach provides strong theory: in the case of two-player Zero-Sum games, it will converge to the mixed strategies of the game’s Nash Equilibrium, and when being non-cooperative and mixed at the same time, this approach will provide bounded Regret and will not diverge. The experiments will test this approach against the state of the art algorithms: SPG, NeuRD, ERI-D, and PPO in the standard tasks of Kuhn Poker, Leduc Poker, matrix games, and grid-worlddomains. Overall, this research aims at developing a unified method that combines the principles of evolutionary stability and the efficiency of reinforcement learning to obtain adaptive learning agents that learn to cooperate effectively and efficiently.